Working Clustering
This seems to be an OK way to cluster. The only issue you can run into
when doing it this way is if the servers are spun up within seconds of
each other. A health check on RabbitMQ's port should be able to keep
this issue from becoming an issue though.