Lookup -> Module
Ansible doesn't let you use lookups on a per role basis so I made it
into a module that seems to work a little better anyway since it doesn't
force you to use split() to get the results that you want. Should also
allow for returning more information and making this more friendly for
other roles in the future.